HomeScreen almost finished. studyprogram and course selection work, continue button works, add study programs works, add course not quite functional
It is now possible to select a StudyProgram and a Course on the HomeScreen. On application launch the studyprograms are loaded from the db and the studyprogam menu entries are created. after selecting a study program the corresponding course entries are loaded form the db and the course menu entries are created. if the studyprogram slection changes the course menu entires are reset. the users selection is stored in the new SharedData class.
The continue button forwards only if study program and a course are selected.

study program can be added. courses can not yet be added (db entries are not correct)